requirements.txt
langchain
sentence-transformers
faiss-cpu
rank-bm25
bert-score
pdfplumber
scikit-learn